requirements.txt
# Core web framework - Python 3.12/3.13 compatible versions
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
pydantic-settings==2.6.1

# Database - Latest compatible versions
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis and Caching - Latest versions
redis==5.2.0
hiredis==3.0.1

# Background Tasks - Latest versions  
celery==5.4.0
flower==2.0.1

# AI/ML integrations
openai==1.3.7
transformers==4.36.0
torch==2.1.1

# Image/Video processing
pillow==10.1.0
opencv-python==4.8.1.78
imageio>=2.31.0
moviepy>=1.0.3
numpy>=1.24.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Social Media APIs
python-telegram-bot>=20.7

# Email
email-validator>=2.1.0

# File handling
python-dotenv>=1.0.0
PyYAML>=6.0

# Date/Time
python-dateutil>=2.8.0

# Monitoring and Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0

# Additional utilities
typing-extensions>=4.8.0
click>=8.1.0
rich>=13.7.0

# HTTP client for API integrations
httpx==0.25.2
requests==2.31.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Task queue
celery==5.3.4
redis==5.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Docker and deployment
gunicorn==21.2.0
email-validator>=2.0.0
